[DEV] adjust dependencies by implementing pydeps adjustments (#184)
* dependencies before changing anything

created with: pydeps wahoomc --reverse

* move download_url_to_file to downloader.py

* dependencies after merge

* move write_to_file and unzip to downloader.py

* dependencies after merge
